src/requirements.txt
asgiref==3.5.2
Django==4.1
django-tinymce==3.5.0
Pillow==9.2.0
sqlparse==0.4.2
tzdata==2022.2
